+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## NEXT ()
 
 - [#23](https://github.com/arthurvasconcelos/vue-izitoast/issues/23) [BUGFIX] Fixed undefined toast in Vue prototype.
+- [#6](https://github.com/arthurvasconcelos/vue-izitoast/issues/6) [IMPROVEMENT] Improved error message of `_checkParams` to show correct method signature.
 
 ### Breaking Changes